% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.matrix.spectra}
\alias{as.matrix.spectra}
\title{Convert spectra to matrix}
\usage{
\method{as.matrix}{spectra}(spec, fix_names = "none")
}
\arguments{
\item{spec}{spectra object}

\item{fix_names}{Use make.names to normalize names?
Pick one: "none" "row" "col" "both".}
}
\value{
matrix of spectral reflectance. columns are wavelengths and rows are
        samples
}
\description{
Convert spectra to matrix
}

