% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.spectra.data.frame}
\alias{as.spectra.data.frame}
\title{Convert data.frame to spectra}
\usage{
\method{as.spectra}{data.frame}(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
spectra object
}
\description{
Convert data.frame to spectra
}
\author{
meireles
}

