% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coersion.R
\name{as.spectra.matrix}
\alias{as.spectra.matrix}
\title{Convert matrix to spectra}
\usage{
as.spectra.matrix(x)
}
\arguments{
\item{x}{matrix}
}
\value{
spectra object
}
\description{
Convert matrix to spectra
}

