% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coersion.R
\name{as.spectra}
\alias{as.spectra}
\title{Convert matrix or data frame to spectra}
\usage{
as.spectra(x)
}
\arguments{
\item{x}{matrix or dataframe. See details for format requirements}
}
\value{
spectra object
}
\description{
Convert matrix or data frame to spectra
}

