% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as_spectra.matrix}
\alias{as_spectra.matrix}
\title{Convert matrix to spectra}
\usage{
\method{as_spectra}{matrix}(x, name_idx = NULL, meta_idxs = NULL)
}
\arguments{
\item{x}{matrix}

\item{name_idx}{column index with sample names. Defaults to NULL}

\item{meta_idxs}{column indices with metadata (not name and not value).
Defaults to NULL}
}
\value{
spectra object
}
\description{
Convert matrix to spectra
}
\author{
Jose Eduardo Meireles
}
