% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{as_tibble.spectra}
\alias{as_tibble.spectra}
\title{Convert spectra to a tibble}
\usage{
\method{as_tibble}{spectra}(x, ...)
}
\arguments{
\item{x}{spectra object}

\item{...}{additional arguments passed to \code{\link{to_long}}}
}
\value{
tibble in long format
}
\description{
A tibble version of \code{\link{to_long}}. Requires the \code{tibble}
package.
}
\examples{
library(spectrolab)
spec = as_spectra(spec_matrix_example, name_idx = 1)
if(requireNamespace("tibble", quietly = TRUE)){
    tbl = tibble::as_tibble(spec)
}
}
\author{
Jose Eduardo Meireles
}
