% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{autoplot.spectra}
\alias{autoplot.spectra}
\title{Plot spectra with ggplot2}
\usage{
\method{autoplot}{spectra}(object, ...)
}
\arguments{
\item{object}{spectra object}

\item{...}{additional arguments (not currently used)}
}
\value{
a ggplot object
}
\description{
Draws one line per sample (wavelength on the x axis, value on the y axis).
Requires the \code{ggplot2} package. For base-R plotting, see
\code{\link{plot.spectra}}.
}
\examples{
library(spectrolab)
spec = as_spectra(spec_matrix_example, name_idx = 1)
if(requireNamespace("ggplot2", quietly = TRUE)){
    ggplot2::autoplot(spec)
}
}
\author{
Jose Eduardo Meireles
}
