% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spectra.R
\name{c.spectra}
\alias{c.spectra}
\alias{rbind.spectra}
\title{Combine spectra with c() or rbind()}
\usage{
\method{c}{spectra}(...)

\method{rbind}{spectra}(...)
}
\arguments{
\item{...}{two or more spectra objects}
}
\value{
a combined spectra object
}
\description{
\code{c.spectra} and \code{rbind.spectra} both combine two or more spectra
objects by repeatedly applying \code{\link{combine}}. Without these
methods, \code{c(s1, s2)} silently degrades to a plain \code{list} and
\code{rbind(s1, s2)} silently degrades to a bare \code{matrix} -- both
lose the \code{spectra} class rather than erroring, which is easy to miss.
}
\section{Functions}{
\itemize{
\item \code{rbind(spectra)}: Combine spectra with rbind()

}}
\examples{
library(spectrolab)
s1 = as_spectra(spec_matrix_example, name_idx = 1)[1:2, ]
s2 = as_spectra(spec_matrix_example, name_idx = 1)[3:4, ]
c(s1, s2)
rbind(s1, s2)
}
\author{
Jose Eduardo Meireles
}
