% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spectra.R
\name{combine}
\alias{combine}
\alias{combine.spectra}
\title{Combine spectral datasets}
\usage{
combine(s1, s2)

\method{combine}{spectra}(s1, s2)
}
\arguments{
\item{s1}{spectra object 1}

\item{s2}{spectra object 1}
}
\value{
combined spectra object
}
\description{
\code{combine} binds two spectral datasets. Both spectra must have the
very same wavelenegth labels, but different metadata are acceptable
}
\section{Methods (by class)}{
\itemize{
\item \code{spectra}: Combines two spectral datasets
}}

\author{
Jose Eduardo Meireles
}
