% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuum_removal.R
\name{continuum_removal}
\alias{continuum_removal}
\alias{continuum_removal.spectra}
\title{Continuum removal}
\usage{
continuum_removal(x, ...)

\method{continuum_removal}{spectra}(x, ...)
}
\arguments{
\item{x}{spectra object. bands must be strictly increasing}

\item{...}{additional arguments (not currently used)}
}
\value{
spectra object with continuum-removed spectra
}
\description{
\code{continuum_removal} divides each spectrum by its continuum -- the
upper convex hull over wavelength/value space -- which suppresses broad
absorption features and emphasizes local absorption depth. Values in the
output are no longer reflectance/radiance.
}
\section{Methods (by class)}{
\itemize{
\item \code{continuum_removal(spectra)}: Continuum removal

}}
\examples{
library(spectrolab)
spec = as_spectra(spec_matrix_example, name_idx = 1)
spec_cr = continuum_removal(spec)
}
\author{
Jose Eduardo Meireles
}
