% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{default_spec_regions}
\alias{default_spec_regions}
\title{Return default spectral regions matrix}
\usage{
default_spec_regions()
}
\value{
matrix with default_spec_regions
}
\description{
Return default spectral regions matrix
}
\examples{
library(spectrolab)
# matrix that defines regions on the spectra
# Useful for plotting w/ plot_regions()
}
\author{
Jose Eduardo Meireles
}
