% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savitzky_golay.R
\name{deriv_spectra}
\alias{deriv_spectra}
\title{Compute spectral derivatives}
\usage{
deriv_spectra(x, order = 1, p = 3, n = p + 3 - p\%\%2, quiet = FALSE)
}
\arguments{
\item{x}{spectra object. bands must be strictly increasing}

\item{order}{derivative order. Must be a positive integer. Defaults to 1}

\item{p}{polynomial order. Defaults to 3}

\item{n}{filter length (number of bands in the moving window). Must be odd
and greater than p. Defaults to \code{p + 3 - p \%\% 2}}

\item{quiet}{boolean. Suppress the message about y values no longer being
reflectance/radiance? Defaults to FALSE}
}
\value{
spectra object with derivative spectra. Values are no longer
        reflectance/radiance
}
\description{
\code{deriv_spectra} computes the order-th derivative of each spectrum using
a Savitzky-Golay filter. Requires the \code{signal} package.
}
\examples{
library(spectrolab)
spec = as_spectra(spec_matrix_example, name_idx = 1)
if(requireNamespace("signal", quietly = TRUE)){
    spec_d1 = deriv_spectra(spec, order = 1)
}
}
\author{
Jose Eduardo Meireles
}
