% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overloads.R
\name{dim.spectra}
\alias{dim.spectra}
\title{Get dimension of spectra object}
\usage{
\method{dim}{spectra}(spec)
}
\arguments{
\item{spec}{spectra object}
}
\value{
tuple of integers. c("n_samples", "n_wavelengths")
}
\description{
Get dimension of spectra object
}

