% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{enforce01<-}
\alias{enforce01<-}
\title{Enforce reflectance between 0 and 1}
\usage{
enforce01(x) <- value
}
\arguments{
\item{x}{spectra object}

\item{value}{boolean.}
}
\value{
nothing. has a *side effect* of changing if a constraint is enforced
}
\description{
\code{enforce01<-} sets or unsets a spectra reflectance constraint (0 - 1)
}
\examples{
library(spectrolab)
spec = as.spectra(spec_matrix_example)

# Force to be between 0 and 1.
enforce01(spec) = TRUE

# So this operation will throw an error
 \dontrun{ reflectance(spec) = reflectance(spec) * 1000 }
}
\author{
Jose Eduardo Meireles
}
