% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{enforce01}
\alias{enforce01}
\alias{enforce01.spectra}
\title{reflectance constraint status}
\usage{
enforce01(x)

\method{enforce01}{spectra}(x)
}
\arguments{
\item{x}{spectra object}
}
\value{
Boolean
}
\description{
\code{enforce01} gets if a reflectance constraint (0 - 1) is being enforced
}
\section{Methods (by class)}{
\itemize{
\item \code{spectra}: Get reflectance constraint status
}}

\examples{
library(spectrolab)
spec = as.spectra(spec_matrix_example, name_idx = 1)
# is reflectance being forced to be between 0 and 1?
enforce01(spec)
}
\author{
Jose Eduardo Meireles
}
