% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wvl_filename.R
\name{filenames.wvl.svc}
\alias{filenames.wvl.svc}
\alias{read.filenames.svc}
\alias{read.wvl.svc}
\title{Read SVC filenames and wavelengths.}
\usage{
read.filenames.svc(inputdir, inclbad = FALSE, inclwr = FALSE)

read.wvl.svc(inputdir)
}
\arguments{
\item{inputdir}{Input directory containing .sig files.
Higher level directories are supported.}

\item{inclbad}{Logical. Should files flagged "bad" be included?}

\item{inclwr}{Logical. Should white reference files, flagged
"WR" be included?}
}
\value{
The output is a dataframe.
}
\description{
Function to read filenames and wavelengths from .sig files
}
\examples{
\dontrun{
read.filenames.svc("path_to_your_folder")
read.wvl.svc("path_to_your_folder")
}
}

