% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wvl_filename.R
\name{filenames_wvl}
\alias{filenames_wvl}
\alias{read_filenames}
\alias{read_wvl}
\title{Read SVC filenames and wavelengths.}
\usage{
read_filenames(inputdir, inclbad = FALSE, inclwr = FALSE)

read_wvl(inputdir)
}
\arguments{
\item{inputdir}{Input directory containing .sig files.
Higher level directories are supported.}

\item{inclbad}{Logical. Should files flagged "bad" be included?}

\item{inclwr}{Logical. Should white reference files, flagged
"WR" be included?}
}
\value{
The output is a dataframe.
}
\description{
Function to read filenames and wavelengths from .sig files
}
\examples{
\dontrun{
read_filenames("path_to_your_folder")
read_wvl("path_to_your_folder")
}
}

