% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{fwhm_from_band_diff}
\alias{fwhm_from_band_diff}
\title{Get the FWHM from the difference between band values}
\usage{
fwhm_from_band_diff(bands)
}
\arguments{
\item{bands}{band values. numeric}
}
\value{
FWHM as a numeric vector
}
\description{
Get the FWHM from the difference between band values
}
\keyword{internal}
