% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_operator_overloads.R
\name{\%*\%.default}
\alias{\%*\%.default}
\title{Matrix multiplication}
\usage{
\method{\%*\%}{default}()
}
\arguments{
\item{x}{matrix 1}

\item{y}{matrix 2}
}
\value{
matrix or vector
}
\description{
Default matrix multiplication operator
}

