% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuum_removal.R
\name{i_continuum_remove_one}
\alias{i_continuum_remove_one}
\title{Remove the continuum from a single spectrum}
\usage{
i_continuum_remove_one(r, w)
}
\arguments{
\item{r}{numeric vector of reflectance/radiance values}

\item{w}{numeric vector of wavelengths, strictly increasing, same length as r}
}
\value{
numeric vector, r divided by its continuum
}
\description{
Remove the continuum from a single spectrum
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
