% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_overlap.R
\name{i_find_sensor_bounds}
\alias{i_find_sensor_bounds}
\title{Find sensor bounds}
\usage{
i_find_sensor_bounds(x, idx = TRUE)
}
\arguments{
\item{x}{wavelength vector}

\item{idx}{boolean. return indices? defaults to TRUE}
}
\value{
data.frame with sensor bounds
}
\description{
\code{i_find_sensor_bounds} finds the bounds bwteen sen
}
\author{
Jose Eduardo Meireles
}
