% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_sensors.R
\name{i_find_sensor_overlap_bounds}
\alias{i_find_sensor_overlap_bounds}
\title{Find sensor overlap bounds}
\usage{
i_find_sensor_overlap_bounds(x, idx = TRUE)
}
\arguments{
\item{x}{band vector}

\item{idx}{boolean. return indices? defaults to TRUE}
}
\value{
data.frame with sensor bounds
}
\description{
\code{i_find_sensor_overlap_bounds} finds the overlap bounds between sensors
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
