% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{i_gain_correction}
\alias{i_gain_correction}
\title{Compute the corrected values for the adjusted side of one junction (one sample)}
\usage{
i_gain_correction(
  bands_fixed,
  vals_fixed,
  bands_adj,
  vals_adj,
  w,
  config,
  fixed_side
)
}
\arguments{
\item{bands_fixed, vals_fixed}{fixed-side bands and values}

\item{bands_adj, vals_adj}{adjusted-side bands and values}

\item{w}{splice wavelength}

\item{config}{splice_config}

\item{fixed_side}{"left" or "right"}
}
\value{
corrected \code{vals_adj}
}
\description{
Compute the corrected values for the adjusted side of one junction (one sample)
}
\keyword{internal}
