% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graveyard.R
\name{i_index_pairwise_combn}
\alias{i_index_pairwise_combn}
\title{Pairwise index combinations}
\usage{
i_index_pairwise_combn(n)
}
\arguments{
\item{n}{number of indices}
}
\value{
list of matrices. For each focal index up to n - 1, a matrix with the
other indexes it interacts with (2nd col) and the "absolute position" of that
pairwise interaction (1st column) is returned.
}
\description{
\code{i_index_pairwise_combn} returns the split pairwise combination of idxs
in a matrix
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
