% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{i_is_increasing}
\alias{i_is_increasing}
\title{Is increasing}
\usage{
i_is_increasing(x)
}
\arguments{
\item{x}{bands}
}
\value{
boolean
}
\description{
\code{i_is_increasing} tests if numerical values are increasing
}
\details{
Many transform functions can only (or at least should only) be applied to
spectra with increasing band values. \code{i_is_increasing} tests for that.
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
