% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{i_is_whole}
\alias{i_is_whole}
\title{Test if argument is a whole number}
\usage{
i_is_whole(x)
}
\arguments{
\item{x}{single value or vector of numbers}
}
\value{
boolean
}
\description{
Test if argument is a whole number
}

