% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{i_match_label}
\alias{i_match_label}
\title{Match label}
\usage{
i_match_label(
  x,
  i,
  full = FALSE,
  allow_empty_lookup = FALSE,
  allow_negative = FALSE
)
}
\arguments{
\item{x}{label vector}

\item{i}{picked label or NULL}

\item{full}{boolean. If TRUE, a full list of results is returned}

\item{allow_empty_lookup}{boolean. If TRUE, x is allowed to be NULL. Defaults
to false}

\item{allow_negative}{boolean. Allow labels to be negative? Defaults to FALSE}
}
\value{
matched indices, or list in case full = TRUE
}
\description{
Match label
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
