% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{i_match_label_or_idx}
\alias{i_match_label_or_idx}
\title{Match label or index}
\usage{
i_match_label_or_idx(
  x,
  i,
  full = FALSE,
  allow_empty_lookup = FALSE,
  allow_negative = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x}{label vector}

\item{i}{picked label or idx or NULL}

\item{full}{boolean. If TRUE, a full list of results is returned}

\item{allow_empty_lookup}{boolean. If TRUE, x is allowed to be NULL. Defaults
to false}

\item{allow_negative}{boolean. Allow indices to be negative? Defaults to FALSE}

\item{verbose}{boolean. Emit diagnostic messages while resolving the fallback
label-vs-index match? Defaults to FALSE (quiet in normal pipelines).}
}
\value{
matched indices
}
\description{
Match label or index
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
