% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_overlap.R
\name{i_match_overlap_svc}
\alias{i_match_overlap_svc}
\title{TODO}
\usage{
i_match_overlap_svc(x, cut_points)
}
\arguments{
\item{x}{TODO}

\item{cut_points}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
