% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{i_mav}
\alias{i_mav}
\title{Moving Average}
\usage{
i_mav(x, n = 3, sides = 2)
}
\arguments{
\item{x}{numeric vector}

\item{n}{number of points going into the average}

\item{sides}{TODO}
}
\value{
numeric vector
}
\description{
\code{i_mav} computes the moving average of a vector.
}
\references{
http://stackoverflow.com/questions/743812/calculating-moving-average-in-r
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
