% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{i_meta}
\alias{i_meta}
\title{Internal metadata constructor for spectra}
\usage{
i_meta(x, nsample, allow_null = TRUE, match_nsample = FALSE)
}
\arguments{
\item{x}{data.frame}

\item{nsample}{number of samples in spectra}

\item{allow_null}{boolean. If TRUE (default) and x is NULL, the function will
return NULL regardless of nsample}

\item{match_nsample}{extend or trim the metadata to match the number of samples?}
}
\value{
data.frame
}
\description{
\code{i_meta} constructs a metadata data.frame in the appropriate format
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
