% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{i_mind_the_gap_smoothing}
\alias{i_mind_the_gap_smoothing}
\title{Warn if gap between bands is too wide}
\usage{
i_mind_the_gap_smoothing(x)
}
\arguments{
\item{x}{spectra}
}
\value{
nothing. warn if gap is too wide
}
\description{
Warn if gap between bands is too wide
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
