% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{i_nearest_band_idx}
\alias{i_nearest_band_idx}
\title{Find the nearest band to a target wavelength}
\usage{
i_nearest_band_idx(b, target, tolerance)
}
\arguments{
\item{b}{numeric vector of band wavelengths}

\item{target}{target wavelength}

\item{tolerance}{maximum allowed distance (in the same units as b) between
target and the nearest available band}
}
\value{
integer index into b
}
\description{
Find the nearest band to a target wavelength
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
