% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{i_plot_boundaries}
\alias{i_plot_boundaries}
\title{Get plot boundaries in user sapce}
\usage{
i_plot_boundaries(return_mat = FALSE)
}
\arguments{
\item{return_mat}{boolean. return a matrix? defaults to FALSE}
}
\value{
vector or matrix, depending on return_mat value
}
\description{
Get plot boundaries in user sapce
}

