% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{i_ramp_weight}
\alias{i_ramp_weight}
\title{Ramp weight: 0 at the low edge, 1 at the high edge of the window}
\usage{
i_ramp_weight(xx, x0, x1, shape)
}
\description{
Ramp weight: 0 at the low edge, 1 at the high edge of the window
}
\keyword{internal}
