% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectra.R
\name{i_read_ascii_metadata}
\alias{i_read_ascii_metadata}
\title{Read metadata}
\usage{
i_read_ascii_metadata(
  file_paths,
  sample_type,
  max_lines,
  sep_char,
  meta_tags,
  tag_sep
)
}
\arguments{
\item{file_paths}{paths}

\item{sample_type}{target or reference}

\item{max_lines}{max number of lines to read}

\item{sep_char}{separator of data within a field}

\item{meta_tags}{tags that match the metadata fields in the file}

\item{tag_sep}{char that separates the tags from the data in the file}
}
\description{
Read metadata
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
