% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectra.R
\name{i_read_asd_spectra}
\alias{i_read_asd_spectra}
\title{Parser for ASD's `.asd`}
\usage{
i_read_asd_spectra(file_paths, type = "target_reflectance", divide_refl_by)
}
\arguments{
\item{file_paths}{paths for files already parsed by `read_spectra`}

\item{type}{Data type to read. "target_refl", "target_rad", "reference_rad".
Defaults to "target_refl".}

\item{divide_refl_by}{divide values by this}

\item{format}{choice of ASD format. Either "binary" or "txt"}
}
\value{
spectra object
}
\description{
Parser for ASD's `.asd`
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
