% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_sensors.R
\name{i_remove_duplicated_band}
\alias{i_remove_duplicated_band}
\title{Remove duplicated band}
\usage{
i_remove_duplicated_band(x, boundary)
}
\arguments{
\item{x}{spectra object}

\item{boundary}{boundary (double). If < wvl value, the first dup is pruned}
}
\value{
spectra object
}
\description{
\code{i_remove_duplicated_band} removes a duplicated band
}
\details{
SVC instruments usually have a certain band that is repeated in the
overlap region of different sensors, e.g. 1005.5 nm. The choice of which
duplicate band to remove (1005.5 in sensor 1 or sensor 2) will depend
on boundary. If the duplicated value is greater than boundary, the first
sensor duplicate is removed. Else, the second dup is removed.
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
