% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_overlap.R
\name{i_remove_duplicated_wavelength}
\alias{i_remove_duplicated_wavelength}
\title{Remove duplicated wavelength}
\usage{
i_remove_duplicated_wavelength(x, boundary)
}
\arguments{
\item{x}{spectra object}

\item{boundary}{sensor boundary}
}
\value{
spectra object
}
\description{
Remove duplicated wavelength
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
