% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{i_report_validation}
\alias{i_report_validation}
\title{Report the result of a validation check (internal)}
\usage{
i_report_validation(msg, stop)
}
\arguments{
\item{msg}{character vector of problem messages (empty if valid)}

\item{stop}{boolean. Error (TRUE) or warn (FALSE) when problems are found?}
}
\value{
invisibly TRUE if valid, FALSE otherwise
}
\description{
Report the result of a validation check (internal)
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
