% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spectra.R
\name{i_smooth_mav_spectra}
\alias{i_smooth_mav_spectra}
\title{Smooth moving average for spectra}
\usage{
i_smooth_mav_spectra(x, n = NULL, save_wvls_to_meta = TRUE)
}
\arguments{
\item{x}{spectra object}

\item{n}{= NULL}

\item{save_wvls_to_meta}{boolean. keep lost ends of original wvls in metadata}
}
\value{
spectra object
}
\description{
Smooth moving average for spectra
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
