% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spectra.R
\name{i_smooth_spline_spectra}
\alias{i_smooth_spline_spectra}
\title{Smooth spline functions for spectra}
\usage{
i_smooth_spline_spectra(x, parallel = TRUE, ...)
}
\arguments{
\item{x}{spectra object. Wavelengths must be strictly increasing}

\item{parallel}{boolean. Do computation in parallel? Defaults to TRUE.
Unfortunately, the parallelization does not work on Windows.}

\item{...}{additional parameters passed to smooth.spline except nknots, which
is computed internally}
}
\value{
a list of spline functions
}
\description{
Gets spline functions for each spectrum in a spectra object.
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
