% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spectra.R
\name{i_test_increasing_wavelengths}
\alias{i_test_increasing_wavelengths}
\title{Are wavelengths increasing}
\usage{
i_test_increasing_wavelengths(x, stop = TRUE, call = FALSE)
}
\arguments{
\item{x}{wavelengths}

\item{stop}{boolean. Throw error if test fails? Defaults to TRUE}

\item{call}{boolean. If stop = TRUE, should the function call be printed?}
}
\value{
boolean
}
\description{
Many transform functions can only (or at least should only) be applied to
spectra with monotonically varying, very likely increasing) wavelength values.
\code{i_test_increasing_wavelengths} tests that case and may throw an error
or return the boolen result from the test.
}
\author{
meireles
}

