% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuum_removal.R
\name{i_upper_hull_idx}
\alias{i_upper_hull_idx}
\title{Find the upper convex hull ("continuum") of a spectrum}
\usage{
i_upper_hull_idx(w, r)
}
\arguments{
\item{w}{numeric vector of wavelengths, strictly increasing}

\item{r}{numeric vector of reflectance/radiance values, same length as w}
}
\value{
integer vector of indices (into w/r), sorted by wavelength
}
\description{
\code{i_upper_hull_idx} finds the indices of \code{r} that lie on the upper
convex hull of the (w, r) point set, i.e. the continuum line. A point of the
convex hull is on the upper hull if it lies on or above the chord connecting
the leftmost and rightmost points -- points on the lower hull lie below it.
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
