% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{i_value}
\alias{i_value}
\title{Internal constructor for value matrix}
\usage{
i_value(x, nwavelengths = NULL, nsample = NULL)
}
\arguments{
\item{x}{numeric matrix, dataframe or vector (in case of single spectrum)}

\item{nwavelengths}{Integer of expected number of wavelengths.
If NULL (default) checking is skipped.}

\item{nsample}{Integer of expected number of samples.
If NULL (default) checking is skipped.}
}
\value{
data conformable to relative value: numeric matrix
}
\description{
\code{i_value} constructs value matrix in the appropriate format
}
\details{
Coerces input form different formats into data conformable to value,
which is a numeric matrix with no dimension names.
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
