% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{i_wavelengths}
\alias{i_wavelengths}
\title{Construct wavelength names in the appropriate format}
\usage{
i_wavelengths(x, nwavelengths = NULL)
}
\arguments{
\item{x}{vector of wavelengths. numeric or character}

\item{nwavelengths}{Integer of expected number of wavelengths.
If NULL (default) checking is skipped.}
}
\value{
vector of wavelengths
}
\description{
Construct wavelength names in the appropriate format
}

