% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{is_spectra}
\alias{is_spectra}
\title{Is it a spectra object?}
\usage{
is_spectra(x)
}
\arguments{
\item{x}{any object}
}
\value{
boolean
}
\description{
\code{is_spectra} tests if the argument is a spectra class object
}
\examples{
library(spectrolab)
spec  = as.spectra(spec_matrix_example, name_idx = 1)
spec1 = unclass(spec)
is_spectra(spec)
is_spectra(spec1)
}
\author{
Jose Eduardo Meireles
}
