% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpcorr_SVC.R
\name{jump_corr}
\alias{jump_corr}
\title{Jump corrector for SVC data}
\usage{
jump_corr(spec)
}
\arguments{
\item{spec}{spectra object with sensor jump between VIS/NIR and NIR/SWIR}
}
\value{
spectra object: jump corrected and interpolated to 1nm resolution
}
\description{
Jump corrector for SVC data
}
\examples{

\dontrun{
jump_corr("raw_spectra_Acer")
}

}

