% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_overlap.R
\name{match_overlap}
\alias{match_overlap}
\alias{match_overlap.spectra}
\title{Match spectra at sensor overlap regions}
\usage{
match_overlap(x, cut_points = c(990, 1900))

\method{match_overlap}{spectra}(x, cut_points = c(990, 1900))
}
\arguments{
\item{x}{spectra object}

\item{cut_points}{cut points for transition between sensors. must be of length
number of sensors - 1. Defaults to c(990, 1900)}
}
\value{
spectra object
}
\description{
Match spectra at sensor overlap regions
}
\section{Methods (by class)}{
\itemize{
\item \code{spectra}: Match sensor ovelap regions
}}
\author{
Jose Eduardo Meireles
}

