% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_overlap.R
\name{match_sensor_transition}
\alias{match_sensor_transition}
\alias{match_sensor_transition.spectra}
\title{Match spectra at sensor transitions}
\usage{
match_sensor_transition(x, instrument = c("asd", "svc", "psr"),
  cut_points = c(990, 1900))

\method{match_sensor_transition}{spectra}(x, instrument = c("asd", "svc",
  "psr"), cut_points = c(990, 1900))
}
\arguments{
\item{x}{spectra object}

\item{cut_points}{cut points defining transition between sensors. must be of
length number of sensors - 1. Defaults to c(990, 1900)}
}
\value{
spectra object
}
\description{
Match spectra at sensor transitions
}
\section{Methods (by class)}{
\itemize{
\item \code{spectra}: Match sensor ovelap regions
}}

\author{
Jose Eduardo Meireles
}
