% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_overlap.R
\name{match_transition}
\alias{match_transition}
\title{Match spectra at sensor transitions}
\usage{
match_transition(x, splice_at, interpolate_wvl = 5)
}
\arguments{
\item{x}{spectra object}

\item{splice_at}{wavelengths that serve as splice poits. Typically the
beginings of sensor 2 and sensor 3.}
}
\value{
spectra object
}
\description{
Match spectra at sensor transitions
}
\author{
Jose Eduardo Meireles
}
