% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_summary_overloads.R
\name{max.spectra}
\alias{max.spectra}
\title{Maximum value}
\usage{
\method{max}{spectra}(..., na.rm = FALSE)
}
\arguments{
\item{...}{spectra object}

\item{na.rm}{boolean. remove NAs? Defaults to FALSE}
}
\value{
single numeric value
}
\description{
\code{max} Returns the maximum value value in a spectra object
}
\examples{
library(spectrolab)
spec = as.spectra(spec_matrix_example, name_idx = 1)
max(spec)
}
\author{
Jose Eduardo Meireles
}
