% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spectra.R
\name{mean.spectra}
\alias{mean.spectra}
\title{Mean spectrum}
\usage{
\method{mean}{spectra}(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{spectra}

\item{na.rm}{boolean. remove NAs?}

\item{...}{nothing}
}
\value{
single spectrum
}
\description{
Mean spectrum
}
\author{
Jose Eduardo Meireles
}
