% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spectra.R
\name{median.spectra}
\alias{median.spectra}
\title{Median spectrum}
\usage{
\method{median}{spectra}(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{spectra}

\item{na.rm}{boolean. remove NAs?}

\item{...}{nothing}
}
\value{
single spectrum
}
\description{
Median spectrum
}
\author{
Jose Eduardo Meireles
}
