% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{meta<-}
\alias{meta<-}
\title{Set metadata}
\usage{
meta(x, label = NULL, sample = NULL) <- value
}
\arguments{
\item{x}{spectra object (lhs)}

\item{label}{metadata column label}

\item{sample}{sample name}

\item{value}{rhs. TODO}
}
\value{
the modified `spectra` object
}
\description{
\code{meta} sets metadata
}
\examples{
library(spectrolab)
spec = as_spectra(spec_matrix_example, name_idx = 1)
meta(spec, "random") = rnorm(nrow(spec), mean(10), sd = 2)
}
\author{
Jose Eduardo Meireles
}
