% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_summary_overloads.R
\name{min.spectra}
\alias{min.spectra}
\title{Minimum value}
\usage{
\method{min}{spectra}(..., na.rm = FALSE)
}
\arguments{
\item{...}{spectra object}

\item{na.rm}{boolean. remove NAs? Defaults to FALSE}
}
\value{
single numeric value
}
\description{
\code{min} Returns the minimum  value in a spectra object
}
\examples{
library(spectrolab)
spec = as_spectra(spec_matrix_example, name_idx = 1)
min(spec)
}
\author{
Jose Eduardo Meireles
}
