% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{names.spectra}
\alias{names.spectra}
\title{Get sample names from spectra}
\usage{
\method{names}{spectra}(spec)
}
\arguments{
\item{spec}{A spectra object}
}
\value{
vector of sample names
}
\description{
Get sample names from spectra
}

