% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spectra.R
\name{normalize}
\alias{normalize}
\alias{normalize.spectra}
\title{Vector normalize spectra}
\usage{
normalize(x, ...)

\method{normalize}{spectra}(x, ...)
}
\arguments{
\item{x}{spectra object to be vector normalized}

\item{...}{nothing}
}
\value{
spectra object with normalized spectra
}
\description{
\code{normalize} returns a spectra obj with vector normalized reflectances
}
\section{Methods (by class)}{
\itemize{
\item \code{spectra}: Vector normalize spectra
}}

