% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_spectra}
\alias{normalize_spectra}
\title{Vector normalization of spectra}
\usage{
normalize_spectra(spec)
}
\arguments{
\item{spec}{spectra object to be vector normalized}
}
\value{
spectra object with normalized spectra
}
\description{
Vector normalization of spectra
}

