% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graveyard.R
\name{pairwise_ratio}
\alias{pairwise_ratio}
\title{Pairwise value ratios}
\usage{
pairwise_ratio(x, simplify = FALSE)
}
\arguments{
\item{x}{spectra}

\item{simplify}{coerce to matrix or keep result as list}
}
\value{
list or matrix
}
\description{
\code{pairwise_ratio} computes pairwise ratios between bands
}
\examples{
library(spectrolab)

# Ratios of visible part of the spectrum
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
