% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectra.R
\name{plot.spectra.svc}
\alias{plot.spectra.svc}
\title{Plot SVC spectra}
\usage{
\method{plot}{spectra.svc}(dat, first, last)
}
\arguments{
\item{dat}{A data.frame, as produced by read.svc. The measurements
need to be per row. The filename, usually the first column should be
exluded to avoid error message.}

\item{first}{Numeric. The first spectrum to plot. Corresponding to a row in a data.frame.}

\item{last}{Numeric. The last spectrum to plot. Corresponding to a row in a data.frame.}
}
\value{
The output is a plot.
}
\description{
Function to plot SVC spectra from a data.frame
}
\examples{
\dontrun{
plot.svc(myspectra[,-1], 1, 10)
}
}

