% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_interactive}
\alias{plot_interactive}
\title{Interactive spectra plot}
\usage{
plot_interactive(spec, colpalette = function(n) RColorBrewer::brewer.pal(n,
  "Dark2"), ...)
}
\arguments{
\item{spec}{spectra object}

\item{colpalette}{a color palette function, e.g. rainbow, terrain.colors, or a
function returned by colorRampPalette() or colorRamps package}

\item{...}{Other arguments passed to plot}
}
\value{
interactive plot
}
\description{
\code{plot_interact} iteratively plots a defined number of spectra. This is
helpful for data inspection and large datasets.
}
\examples{
\dontrun{
# Create a spectra object
spec = as.spectra(spec_matrix_example)

# Start interactive plot
plot_interactive(spec)
}
}
\author{
Anna K. Schweiger and Jose Eduardo Meireles
}
