% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw.R
\name{plot_rawspec}
\alias{plot_rawspec}
\title{Plot unprocessed spectra}
\usage{
plot_rawspec(dat, first, last)
}
\arguments{
\item{dat}{A data.frame, e.g. as produced by read_spec. The measurements
need to be per row. The filename, usually the first column should be
exluded to avoid warning.}

\item{first}{Numeric. The first spectrum to plot. Corresponding to a row in a data.frame.}

\item{last}{Numeric. The last spectrum to plot. Corresponding to a row in a data.frame.}
}
\value{
The output is a plot.
}
\description{
Function to plot unprocessed SVC spectra from a data.frame
}
\examples{
\dontrun{
plot_rawspec(Acer_spectra[,-1], 1, 10)
}
}

