% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_regions}
\alias{plot_regions}
\title{Plot polygons for spectral regions}
\usage{
plot_regions(spec, regions, col = rgb(0.7, 0.7, 0.7, 0.3), border = FALSE,
  add = TRUE, names = TRUE, ...)
}
\arguments{
\item{spec}{spectra object}

\item{regions}{matrix with spectral regions in columns and only two rows named
"begin" and "end". Values are the wavelengths where a spectral
regions begins and ends. See details for an example}

\item{col}{color for regions. single value or vector of length ncol(regions)}

\item{border}{color for region borders. Defaults to FALSE (no border)}

\item{add}{boolean. If TRUE (default) adds polygons to current plot (if a plot
exists) or throws an error if a plot doesn't exist.
If FALSE, a new plot is created **without** any spectra.}

\item{names}{boolean. add region column names on top of the polygons?}

\item{...}{additional parameters passed to polygon()}
}
\value{
nothing. called for its side effect
}
\description{
default spectral regions
spec_regions = cbind("VIS"   = c(begin = 400,  end = 700),
                     "NIR"   = c(begin = 800,  end = 1300),
                     "SWIR1" = c(begin = 1550, end = 1800),
                     "SWIR2" = c(begin = 2000, end = 2400))
}
\author{
Jose Eduardo Meireles
}

