% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_SVC.R
\name{preprocess}
\alias{excl_dip}
\alias{excl_high}
\alias{excl_low}
\alias{excl_shoulder}
\alias{preprocess}
\title{Preprocess spectra}
\usage{
excl_shoulder(spec, refl_high, refl_low)

excl_high(spec, refl_high, wvl_high)

excl_low(spec, refl_low, wvl_low)

excl_dip(spec, wvl_low, wvl_high, refl_diff)
}
\arguments{
\item{spec}{spectra object}

\item{refl_high}{reflectance upper limit}

\item{refl_low}{reflectance lower limit}

\item{wvl_high}{wavelength upper limit}

\item{wvl_low}{wavelength lower limit}

\item{refl_diff}{maximum reflectance difference}
}
\value{
spectra object
}
\description{
Functions to clean up spectral datasets.
}
\details{
The summary function \code{excl_shoulder} uses the reflectance value
at the NIR shoulder (780 nm) and excludes spectra below \code{refl_low} and
above \code{refl_high}.

The functions \code{excl_high} and
\code{refl_high} exclude spectra with reflectance values above or below the
defined limits \code{excl_high} and \code{excl_low} at user defined
wavelengths \code{wvl_high} and \code{wvl_low}, respectively. \code{excl_dip}
exclude spectra if the difference between \code{wvl_high} and \code{wvl_low}
exceeds \code{refl_diff}.
}
\examples{
\dontrun{
excl_shoulder(acer_spectra, refl_high=0.7, refl_low=0.3)
}

\dontrun{
excl_low(acer_spectra, refl_low=0.01, wvl_low=401)
}

\dontrun{
excl_low(acer_spectra, wvl_low=770, wvl_high=800, refl_diff=0.02)
}

}

