% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_summary_overloads.R
\name{print.spectra}
\alias{print.spectra}
\title{Print spectra}
\usage{
\method{print}{spectra}(x, ...)
}
\arguments{
\item{x}{spectra object}

\item{...}{other arguments passed to print. not implemented for spectra}
}
\value{
`x`, invisibly (called for its side effect of printing)
}
\description{
\code{print} prints basic information about the spectra obj to the console
}
\examples{
library(spectrolab)
spec = as_spectra(spec_matrix_example, name_idx = 1)
print(spec)
## or simply
spec
}
\author{
Jose Eduardo Meireles
}
