% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_summary_overloads.R
\name{range.spectra}
\alias{range.spectra}
\title{Range of value values}
\usage{
\method{range}{spectra}(..., na.rm = FALSE)
}
\arguments{
\item{...}{spectra object}

\item{na.rm}{boolean. remove NAs? Defaults to FALSE}
}
\value{
tuple of numeric values (min, max)
}
\description{
\code{range} Returns the range of (min, max) value values in spectra
}
\examples{

library(spectrolab)
spec = as.spectra(spec_matrix_example, name_idx = 1)
range(spec)
}
\author{
Jose Eduardo Meireles
}
