% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ratio}
\alias{ratio}
\title{Pairwise reflectance ratios}
\usage{
ratio(x, simplify = FALSE)
}
\arguments{
\item{x}{spectra}

\item{simplify}{coerce to matrix or keep result as list}
}
\value{
list or matrix
}
\description{
Pairwise reflectance ratios
}
\author{
Jose Eduardo Meireles
}
