% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_svc.R
\name{read_spec}
\alias{read_spec}
\title{Read SVC files}
\usage{
read_spec(inputdir, inclbad = FALSE, inclwr = FALSE, read = FALSE,
  filename = TRUE)
}
\arguments{
\item{inputdir}{Input directory containing .sig files.
Higher level directories are supported.}

\item{inclbad}{Logical. Should files flagged "bad" be included?}

\item{inclwr}{Logical. Should white reference files, flagged
"WR" be included?}

\item{read}{Logical. If TRUE files are read and combined in a data.frame.
If FALSE the filenames are returned.}

\item{filename}{Logical. Should the filename be included as the first column?}
}
\value{
The output is a dataframe.
}
\description{
Function to read .sig files and combine them in a data.frame
}
\examples{
\dontrun{
read_spec("path_to_your_folder")
}
}

