% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{reflectance}
\alias{reflectance}
\title{Get spectra reflectance}
\usage{
reflectance(x)
}
\arguments{
\item{x}{spectra object}
}
\value{
matrix with samples in rows and wavelengths in columns
}
\description{
\code{reflectance} returns the reflectance matrix from spectra
}

