% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample_spec_fwhm}
\alias{resample_spec_fwhm}
\title{Resample spectra}
\usage{
resample_spec_fwhm(spec, new_bands, fwhm)
}
\arguments{
\item{spec}{spectra object}

\item{new_bands}{band values to resample the spectra to}

\item{fwhm}{FWHM for the new bands}
}
\value{
resampled spectra
}
\description{
Resample spectra
}
