% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{sample_names<-}
\alias{sample_names<-}
\title{Set sample names in spectra}
\usage{
sample_names(spec) <- value
}
\arguments{
\item{spec}{Spectra object}
}
\description{
Set sample names in spectra
}

