% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_function_overloads.R
\name{sd}
\alias{sd}
\title{Standard deviation}
\usage{
sd(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an R object which is coercible to one by as.double(x)}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
standard deviation
}
\description{
\code{sd} computes the standard deviation spectrum. Note that values will not
reflect reflectance anymore, but the sd of the reflectance instead.
}
