% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_function_overloads.R
\name{sd.spectra}
\alias{sd.spectra}
\title{Standard deviation spectrum}
\usage{
\method{sd}{spectra}(x, na.rm = TRUE)
}
\arguments{
\item{x}{spectra}

\item{na.rm}{boolean. remove NAs?}
}
\value{
single spectrum
}
\description{
Forces keep_txt_meta = TRUE
}
\examples{
library(spectrolab)
spec = as_spectra(spec_matrix_example, name_idx = 1)
sd(spec)
}
\author{
Jose Eduardo Meireles
}
