% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth}
\alias{smooth}
\title{Generic Smoothing function}
\usage{
smooth(x, ...)
}
\arguments{
\item{x}{data to smooth over}

\item{...}{additional arguments}
}
\value{
smoothed data
}
\description{
Generic Smoothing function
}
