% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_fwhm}
\alias{smooth_fwhm}
\title{Smooth spectra with a gaussian model}
\usage{
smooth_fwhm(x, fwhm = NULL)
}
\arguments{
\item{x}{spectra}

\item{fwhm}{Full Width at Half Maximum.}
}
\value{
smoothed spectra
}
\description{
Smooth spectra with a gaussian model
}
