% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_moving_avg}
\alias{smooth_moving_avg}
\title{Smooth moving average for spectra}
\usage{
smooth_moving_avg(x, n = NULL, save_bands_to_meta = TRUE)
}
\arguments{
\item{x}{spectra object}

\item{n}{number of bands going into each moving average. When \code{NULL}
(default) a window is chosen from the band resolution.}

\item{save_bands_to_meta}{boolean. keep lost ends of original wvls in metadata}
}
\value{
spectra object
}
\description{
Smooth moving average for spectra
}
\author{
Jose Eduardo Meireles
}
