% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spec_matrix_example}
\alias{spec_matrix_example}
\title{Example spectral dataset}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 50 rows and 2102 columns.
}
\usage{
spec_matrix_example
}
\description{
Simulated spectral dataset as a matrix. First column hold species names and
the remaining ones the value values. Wavelength labels are given as
column names
}
\author{
Jose Eduardo Meireles
}
\keyword{datasets}
