% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spec_with_jump}
\alias{spec_with_jump}
\title{Example spectra with jump between sensors}
\format{
An object of class \code{spectra} with 1 rows and 2101 columns.
}
\usage{
spec_with_jump
}
\description{
Spectra object of a simulated spectra with "jumps" between sensors. Used to
illustrate the `match_sensors` function
}
\author{
Jose Eduardo Meireles
}
\keyword{datasets}
