% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spectra.R
\name{split.spectra}
\alias{split.spectra}
\title{Split spectra}
\usage{
\method{split}{spectra}(x, f, drop = FALSE, ...)
}
\arguments{
\item{x}{spectra object}

\item{f}{factor vector defining the grouping. Must have length nrow(x)}

\item{drop}{NOT used}

\item{...}{NOT used}
}
\value{
list of spectra
}
\description{
\code{split} a spectra object into a list of spectra according to grouping f.
}
\examples{
library(spectrolab)
spec = as.spectra(spec_matrix_example, name_idx = 1)
spec_list = split(spec, names(spec))
}
\author{
Jose Eduardo Meireles
}
