% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_summary_overloads.R
\name{str.spectra}
\alias{str.spectra}
\title{Structure of the spectra object}
\usage{
\method{str}{spectra}(object, ...)
}
\arguments{
\item{object}{spectra object}

\item{...}{additioal args. not implemented}
}
\value{
prints to console
}
\description{
Structure of the spectra object
}
\author{
Jose Eduardo Meireles
}
