% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spectra.R
\name{subset_by}
\alias{subset_by}
\alias{subset_by.spectra}
\title{Subset spectra by factor}
\usage{
subset_by(x, by, n_max, random = TRUE)

\method{subset_by}{spectra}(x, by, n_max, random = TRUE)
}
\arguments{
\item{x}{spectra object}

\item{by}{vector coercible to factor and of same length as nrow(x)}

\item{n_max}{integer. keep at most this number of spectra per unique `by`}

\item{random}{boolean. Sample randomly or keep first n_max? Defaults to TRUE}
}
\value{
spectra
}
\description{
\code{subset_by} subsets spectra ensuring that each factor `by` appears only
`max` times or less in the spectra dataset.
}
\section{Methods (by class)}{
\itemize{
\item \code{spectra}: Subset spectra by factor
}}

\author{
Jose Eduardo Meireles
}
