% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_operator_overloads.R
\name{t.spectra}
\alias{t.spectra}
\title{Spectra Transpose}
\usage{
\method{t}{spectra}(x)
}
\arguments{
\item{x}{spectra}
}
\value{
nothing. operation not allowed
}
\description{
spectra are not transposable. Transpose the reflectance instead
}
\examples{
library(spectrolab)
s = as.spectra(spec_matrix_example, name_idx = 1)

# This will throw an error
\dontrun{
t(s)
}
# But these options should work
t(reflectance(s))
t(as.matrix(s))
}
\author{
Jose Eduardo Meireles
}
