% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{value<-}
\alias{value<-}
\title{Set spectra value}
\usage{
value(x) <- value
}
\arguments{
\item{x}{spectra object}

\item{value}{value to be assigned to the lhs}
}
\value{
the modified `spectra` object
}
\description{
\code{value<-} Assigns the rhs to the value of the lhs spectra obj
}
\examples{
library(spectrolab)
spec = as_spectra(spec_matrix_example, name_idx = 1)
# scale all refletance values by 2
value(spec) = value(spec) * 2
}
\author{
Jose Eduardo Meireles
}
