% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{value}
\alias{value}
\alias{value.spectra}
\title{Get spectra value}
\usage{
value(x)

\method{value}{spectra}(x)
}
\arguments{
\item{x}{spectra object}
}
\value{
matrix with samples in rows and bands in columns
}
\description{
\code{value} returns the value matrix from spectra
}
\section{Methods (by class)}{
\itemize{
\item \code{spectra}: Get spectra value
}}

\examples{
library(spectrolab)
spec = as_spectra(spec_matrix_example, name_idx = 1)
is.matrix(value(spec))
}
\author{
Jose Eduardo Meireles
}
