% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_function_overloads.R
\name{var}
\alias{var}
\title{Variance}
\usage{
var(x, y = NULL, na.rm = FALSE, use)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame}

\item{y}{NULL (default) or a vector, matrix or data frame with compatible
dimensions to x.}

\item{na.rm}{logical. Should missing values be removed?}

\item{use}{an optional character string giving a method for computing covariances in
the presence of missing values. This must be (an abbreviation of) one of
the strings "everything", "all.obs", "complete.obs", "na.or.complete",
or "pairwise.complete.obs"}
}
\value{
variance
}
\description{
\code{var} computes the variance spectrum. Note that values will not reflect
reflectance anymore, but the variance of the reflectance instead.
}
