% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{wavelength_unit}
\alias{wavelength_unit}
\alias{wavelength_unit.spectra}
\title{Get the wavelength unit of a spectra object's bands}
\usage{
wavelength_unit(x)

\method{wavelength_unit}{spectra}(x)
}
\arguments{
\item{x}{a spectra object}
}
\value{
a single character string
}
\description{
\code{wavelength_unit} returns the unit bands are expressed in --
\code{"nm"} by default (every currently supported instrument reports
nanometers), or whatever was captured/set for \code{x}.
}
\section{Methods (by class)}{
\itemize{
\item \code{wavelength_unit(spectra)}: Get the wavelength unit of a spectra object's bands

}}
\examples{
library(spectrolab)
spec = as_spectra(spec_matrix_example, name_idx = 1)
wavelength_unit(spec)
}
\author{
Jose Eduardo Meireles
}
