% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spectra.R
\name{aggregate.spectra}
\alias{aggregate.spectra}
\title{Aggregate}
\usage{
\method{aggregate}{spectra}(x, by, FUN, FUN_meta = NULL, ...)
}
\arguments{
\item{x}{spectra object}

\item{by}{vector of factors to guide the aggregation}

\item{FUN}{function to be applied}

\item{FUN_meta}{function to be applied to metadata. If NULL (default), same
FUN appied to reflectance is used.}

\item{...}{extra args to FUN}
}
\value{
spectra object
}
\description{
Aggregate
}
\author{
Jose Eduardo Meireles
}
