% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.data.frame.spectra}
\alias{as.data.frame.spectra}
\title{Convert spectra to data.frame}
\usage{
\method{as.data.frame}{spectra}(x, row.names = NULL, optional = FALSE,
  fix_names = "none", ...)
}
\arguments{
\item{x}{spectra object}

\item{row.names}{does nothing}

\item{optional}{does nothing}

\item{fix_names}{Use make.names to normalize names?
Pick one: "none" "row" "col" "both".}

\item{...}{does nothing}
}
\value{
data.frame with: sample_name, metadata (if any) and reflectance.
}
\description{
Returns a data.frame that includes sample names, metadata (if present) and
reflectance data. One advantage over as.matrix, is that the metadata are
returned.
}
\author{
meireles
}

