% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coersion.R
\name{as.matrix.spectra}
\alias{as.matrix.spectra}
\title{Convert spectra to matrix}
\usage{
\method{as.matrix}{spectra}(spec, fix_dimnames = TRUE)
}
\arguments{
\item{spec}{spectra object}

\item{fix_dimnames}{boolean. if true (default), names are normalized with name.names()}
}
\value{
matrix of spectral reflectance. columns are wavelengths and rows are samples
}
\description{
Convert spectra to matrix
}

