% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.matrix.spectra}
\alias{as.matrix.spectra}
\title{Convert spectra to matrix}
\usage{
\method{as.matrix}{spectra}(x, fix_names = "none", ...)
}
\arguments{
\item{x}{spectra object}

\item{fix_names}{Use make.names to normalize names?
Pick one: "none" "row" "col" "both".}

\item{...}{does nothing. Here for compatibility with S3 generics}
}
\value{
matrix of spectral value. columns are bands and rows are
        samples
}
\description{
Convert spectra to matrix
}
\examples{
library(spectrolab)
spec = as_spectra(spec_matrix_example, name_idx = 1)
mat  = as.matrix(spec)
}
\author{
Jose Eduardo Meireles
}
