% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coersion.R
\name{as.spectra.data.frame}
\alias{as.spectra.data.frame}
\title{Convert data.frame to spectra}
\usage{
as.spectra.data.frame(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
spectra object
}
\description{
Convert data.frame to spectra
}

