% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{bands<-}
\alias{bands<-}
\title{Set band labels}
\usage{
bands(x) <- value
}
\arguments{
\item{x}{spectra object (lhs)}

\item{value}{rhs}
}
\value{
nothing. called for its side effect.
}
\description{
\code{bands} sets band labels of lhs to the rhs values
}
\examples{
library(spectrolab)
spec = as_spectra(spec_matrix_example, name_idx = 1)
bands(spec) = bands(spec) / 1000
}
\author{
Jose Eduardo Meireles
}
