% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{bands}
\alias{bands}
\alias{bands.spectra}
\title{Get spectra band labels}
\usage{
bands(x, min = NULL, max = NULL, return_num = TRUE)

\method{bands}{spectra}(x, min = NULL, max = NULL, return_num = TRUE)
}
\arguments{
\item{x}{spectra object}

\item{min}{= NULL}

\item{max}{= NULL}

\item{return_num}{boolean. return vector of numeric values (default).
otherwise, a vector of strings is returned}
}
\value{
vector of bands. numeric if `return_num` = TRUE (default).
}
\description{
\code{bands} returns a vector of band labels from spectra
}
\section{Methods (by class)}{
\itemize{
\item \code{bands(spectra)}: Get spectra band labels

}}
\examples{
library(spectrolab)
spec = as_spectra(spec_matrix_example, name_idx = 1)
head(bands(spec))
}
\author{
Jose Eduardo Meireles
}
