% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{default_spec_regions}
\alias{default_spec_regions}
\title{return default spectral regions matrix}
\usage{
default_spec_regions()
}
\value{
matrix with default_spec_regions
}
\description{
return default spectral regions matrix
}
\author{
meireles
}

