% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_summary_overloads.R
\name{dim.spectra}
\alias{dim.spectra}
\title{Get dimension of spectra}
\usage{
\method{dim}{spectra}(x)
}
\arguments{
\item{x}{spectra object}
}
\value{
tuple of integers: c("n_samples", "n_bands")
}
\description{
\code{dim} returns a vector with number of samples and bands (bands)
}
\examples{
library(spectrolab)
spec = as.spectra(spec_matrix_example, name_idx = 1)
dim(spec)
}
\author{
Jose Eduardo Meireles
}
