% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{enforce01<-}
\alias{enforce01<-}
\title{Enforce reflectance between 0 and 1}
\usage{
enforce01(x) <- value
}
\arguments{
\item{x}{spectra object}

\item{value}{boolean.}
}
\value{
nothing. has a *side effect* of changing if a constraint is enforced
}
\description{
\code{enforce01<-} sets or unsets a spectra reflectance constraint (0 - 1)
}
\author{
Jose Eduardo Meireles
}
