% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_operator_overloads.R
\name{\%*\%.spectra}
\alias{\%*\%.spectra}
\title{Matrix multiplication for spectra class}
\usage{
\method{\%*\%}{spectra}(e1, e2)
}
\arguments{
\item{e2}{}
}
\description{
Matrix multiplication for spectra class
}

