% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_sensors.R
\name{guess_splice_at}
\alias{guess_splice_at}
\alias{guess_splice_at.spectra}
\title{Guess splice bands (bounds between senors)}
\usage{
guess_splice_at(x)

\method{guess_splice_at}{spectra}(x)
}
\arguments{
\item{x}{spectra object}
}
\value{
vector of band values
}
\description{
Guess splice bands (bounds between senors)
}
\section{Methods (by class)}{
\itemize{
\item \code{guess_splice_at(spectra)}: Guess splice bands (bounds between senors)

}}
\author{
Jose Eduardo Meireles
}
