% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{has_nir_dip}
\alias{has_nir_dip}
\title{Has a dip in the NIR region?}
\usage{
has_nir_dip(x)
}
\arguments{
\item{x}{spectra object}
}
\value{
boolean vector
}
\description{
\code{has_nir_dip} tests if spectra have a NIR dip
}
\details{
Because an NIR dip is estimated from features of the second derivative, high
frequency noise in the spectra is likely to spoil the algorithm and give false
positives. Make sure that the spectra has been smoothed (`spectrolab::smooth`)
before trying to detect an NIR dip.
}
\author{
Jose Eduardo Meireles
}
