% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{i_bands}
\alias{i_bands}
\title{Internal band constructor for spectra}
\usage{
i_bands(x, nbands = NULL)
}
\arguments{
\item{x}{vector of bands. Either numeric or character}

\item{nbands}{Integer of expected number of bands.
If NULL (default) checking is skipped.}
}
\value{
vector of bands
}
\description{
\code{i_bands} constructs band labels in the appropriate format
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
