% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{i_is_increasing}
\alias{i_is_increasing}
\title{Is increasing}
\usage{
i_is_increasing(x, stop = TRUE, call = FALSE)
}
\arguments{
\item{x}{wavelengths}

\item{stop}{boolean. Throw error if test fails? Defaults to TRUE}

\item{call}{boolean. If stop = TRUE, should the function call be printed?}
}
\value{
boolean
}
\description{
\code{i_is_increasing} tests if wavelength values are increasing
}
\details{
Many transform functions can only (or at least should only) be applied to
spectra with monotonically varying, very likely increasing) wavelength values.
\code{i_is_increasing} tests that case and may throw an error
or return the boolen result from the test.
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
