% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{i_is_index}
\alias{i_is_index}
\title{Compatible with being an index?}
\usage{
i_is_index(x, max_length, quiet = TRUE)
}
\arguments{
\item{x}{numeric values}

\item{max_length}{max acceptable values for x}

\item{quiet}{get warnings?}
}
\value{
boolean
}
\description{
\code{i_is_index} Tests if x fit the requirements of being indices
}

