% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{i_is_index}
\alias{i_is_index}
\title{Compatible with being an index?}
\usage{
i_is_index(x, max_value, allow_negative = FALSE)
}
\arguments{
\item{x}{numeric values}

\item{allow_negative}{boolean. Count negative integers as indices? defaults to FALSE}

\item{max_length}{Max acceptable values for x (inclusive). Must be greater than 1}
}
\value{
boolean
}
\description{
\code{i_is_index} Tests if x fit the requirements of being indices
}
\details{
This function potentially allows negative indices, given that they may be used
with the intent of removing an entry that corresponds to the index. Conversely,
zero is never used as an index in R and is not recognized as such here.
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
