% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{i_is_index}
\alias{i_is_index}
\title{Is number compatible with being an index?}
\usage{
i_is_index(x, max_length, verbose = FALSE)
}
\arguments{
\item{x}{numeric values}

\item{max_length}{max acceptable values for x}

\item{verbose}{get warnings?}
}
\description{
Is number compatible with being an index?
}

