% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{i_is_index}
\alias{i_is_index}
\title{Are numbers compatible with being an index?}
\usage{
i_is_index(x, max_length, quiet = TRUE)
}
\arguments{
\item{x}{numeric values}

\item{max_length}{max acceptable values for x}

\item{verbose}{get warnings?}
}
\description{
Are numbers compatible with being an index?
}

