% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{i_is_whole}
\alias{i_is_whole}
\title{Is whole number?}
\usage{
i_is_whole(x)
}
\arguments{
\item{x}{single value or vector of numbers}
}
\value{
boolean
}
\description{
\code{i_is_whole} Tests if x is (are) whole numbers
}
\author{
Jose Eduardo Meireles
}

