% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{i_match_ij_spectra}
\alias{i_match_ij_spectra}
\title{Get internal indexes for spectra attributes}
\usage{
i_match_ij_spectra(x, i = NULL, j = NULL)
}
\arguments{
\item{x}{spectra}

\item{i}{sample names or indices or boolean vector}

\item{j}{wavelengths or boolean vector, NOT INDICES}
}
\value{
list if row indices and column indices
}
\description{
\code{i_match_ij_spectra} gets index position matching i and j
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
