% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{i_match_ij_spectra}
\alias{i_match_ij_spectra}
\title{Get internal indexes for spectra attributes}
\usage{
i_match_ij_spectra(this, i = NULL, j = NULL)
}
\arguments{
\item{this}{spectra}

\item{i}{sample names or indices}

\item{j}{wavelengths, not indices}
}
\value{
list if row indices and colimn indices
}
\description{
\code{i_match_ij_spectra} gets position matching if the relfectance matrix
}

