% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{i_match_label}
\alias{i_match_label}
\title{Match label}
\usage{
i_match_label(x, i)
}
\arguments{
\item{x}{label vector}

\item{i}{picked label or NULL}
}
\value{
matched indices
}
\description{
Match label
}
\author{
meireles
}

