% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{i_match_label_or_idx}
\alias{i_match_label_or_idx}
\title{Match label or index}
\usage{
i_match_label_or_idx(x, i, full = FALSE)
}
\arguments{
\item{x}{label vector}

\item{i}{picked label or idx or NULL}

\item{full}{boolean. If TRUE, a full list of results is returned}
}
\value{
matched indices
}
\description{
Match label or index
}
\author{
meireles
}

