% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{i_names}
\alias{i_names}
\title{Internal constructor for sample names}
\usage{
i_names(x, nsample = NULL)
}
\arguments{
\item{x}{vector of labels. Should be character.}

\item{nsample}{Integer of expected number of samples.
If NULL (default) checking is skipped.}
}
\value{
vector of sample names coerced to character
}
\description{
\code{i_names} constructs a sample name vector in the appropriate format
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
