% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{i_names}
\alias{i_names}
\title{Internal constructor for sample names}
\usage{
i_names(x, nsample = NULL, prefix = "spec_")
}
\arguments{
\item{x}{vector of labels. Should be character. If numeric, a prefix will be added}

\item{nsample}{Integer of expected number of samples.
If NULL (default) checking is skipped.}

\item{prefix}{String to use as prefix in case an element of x is numeric.
Defaults to "spec_"}
}
\value{
vector of sample names coerced to character
}
\description{
\code{i_names} constructs a sample name vector in the appropriate format
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
