% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{i_names}
\alias{i_names}
\title{Construct sample names vector in the appropriate format}
\usage{
i_names(x, nsample = NULL)
}
\arguments{
\item{x}{vector of labels}

\item{nsample}{Integer of expected number of samples. If NULL (default) checking is skipped.}
}
\value{
vector of sample names
}
\description{
Construct sample names vector in the appropriate format
}

