% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{i_plot_boundaries}
\alias{i_plot_boundaries}
\title{Find plot boundaries in user space}
\usage{
i_plot_boundaries(return_mat = FALSE)
}
\arguments{
\item{return_mat}{return a matrix instead of vector? defaults to FALSE}
}
\value{
vector or matrix, depending on return_mat value
}
\description{
\code{i_plot_boundaries} gets plot boundaries in user space as matrix or vec
}
\author{
Jose Eduardo Meireles
}

