% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectra.R
\name{i_read_ascii_spectra}
\alias{i_read_ascii_spectra}
\title{Internal parser for ASCII format}
\usage{
i_read_ascii_spectra(file_paths, skip_first_n, sep_char, wl_and_refl_cols,
  divide_refl_by, include_white_ref, outside_01_fun, ...)
}
\arguments{
\item{file_paths}{paths for files already parsed by `read_spectra`}

\item{skip_first_n}{skip the first n lines}

\item{sep_char}{separator}

\item{wl_and_refl_cols}{vector of length 2 with index wavelength labels and
reflectance values}

\item{divide_refl_by}{divide reflectance values by this}

\item{include_white_ref}{NOT USED YET, but should read the write reference
from each file}

\item{outside_01_fun}{function to deal with reflectance values outside 0.1.}

\item{...}{NOT USED YET}
}
\value{
single `spectra` file or list of `spectra`
}
\description{
Generic parser for SVC's `.sig` and PSR's `.sed`
}
\author{
Jose Eduardo Meireles
}

