% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectra.R
\name{i_read_ascii_spectra}
\alias{i_read_ascii_spectra}
\title{Internal parser for ASCII format}
\usage{
i_read_ascii_spectra(file_paths, skip_until_tag = NULL, skip_first_n = NULL,
  sep_char, header, wl_col, refl_cols, divide_refl_by, include_white_ref, ...)
}
\arguments{
\item{file_paths}{paths for files already parsed by `read_spectra`}

\item{skip_until_tag}{Tag that precedes the table of reflectance data,
indicating that lines until that tag should be skipped.
Tag is matched with a regexpr.}

\item{skip_first_n}{skip the first n lines. Only used if `skip_until_tag` is
not given}

\item{sep_char}{separator}

\item{header}{boolean. keep header?}

\item{wl_col}{idx or name of wavelength column}

\item{refl_cols}{idx or name of reflectance columns. MULTIPLE}

\item{divide_refl_by}{divide reflectance values by this. MULTIPLE}

\item{include_white_ref}{NOT USED YET, but should read the write reference
from each file}

\item{...}{additional arguments passed to read table}
}
\value{
single `spectra` or list of `spectra`
}
\description{
Generic parser for SVC's `.sig` and PSR's `.sed`
}
\author{
Jose Eduardo Meireles and Anna Schweiger
}
\keyword{internal}
