% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spectra.R
\name{i_smooth_mav_spectra}
\alias{i_smooth_mav_spectra}
\title{Smooth moving average for spectra}
\usage{
i_smooth_mav_spectra(x)
}
\arguments{
\item{x}{spectra object}
}
\value{
spectra object
}
\description{
Smooth moving average for spectra
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
