% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_overlap.R
\name{i_trim_sensor_overlap}
\alias{i_trim_sensor_overlap}
\title{Trim sensor overlap}
\usage{
i_trim_sensor_overlap(x, splice_at)
}
\arguments{
\item{x}{spectra object}

\item{splice_at}{}
}
\description{
Trim sensor overlap
}
