% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_sensors.R
\name{i_trim_sensor_overlap}
\alias{i_trim_sensor_overlap}
\title{Trim sensor overlap}
\usage{
i_trim_sensor_overlap(x, splice_at)
}
\arguments{
\item{x}{spectra object}

\item{splice_at}{wavelengths where to splice sensors. suggests where the
beginning of sensors 2 and 3 should be.}
}
\value{
spectra object
}
\description{
Trim sensor overlap
}
\author{
Jose Eduardo Meireles
}
\keyword{internal}
