% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{i_warn_rejected_gain}
\alias{i_warn_rejected_gain}
\title{Warn that a junction was left uncorrected because its factor was implausible}
\usage{
i_warn_rejected_gain(n_rejected, n_total, w, clamp)
}
\arguments{
\item{n_rejected}{number of samples whose factor failed the check}

\item{n_total}{number of samples}

\item{w}{the junction wavelength}

\item{clamp}{the plausible range}
}
\value{
invisible NULL
}
\description{
Warn that a junction was left uncorrected because its factor was implausible
}
\keyword{internal}
