% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_spectra}
\alias{is_spectra}
\title{Is spectra object?}
\usage{
is_spectra(x)
}
\arguments{
\item{x}{any object}
}
\value{
boolean
}
\description{
\code{is_spectra} tests if the argument is a spectra class object
}

