% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpcorr_SVC.R
\name{jump_corr}
\alias{jump_corr}
\title{Jump corrector for SVC data}
\usage{
jump_corr(path, filename = FALSE)
}
\arguments{
\item{path}{Path to the input directory containing .sig files.
Higher level directories are supported.}

\item{filename}{Logical. Should the filename be included as the first column?}
}
\value{
The output is a dataframe.
}
\description{
Reads .sig files and corrects the jump at the sensor overlap regions,
    leaving the NIR unchanged
}
\examples{
\dontrun{
jump_corr("path_to_your_folder", filename=TRUE)
}
}

