% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_overlap.R
\name{match_sensors}
\alias{match_sensors}
\alias{match_sensors.spectra}
\title{Match spectra at sensor transitions}
\usage{
match_sensors(x, splice_at, interpolate_wvl = 5)

\method{match_sensors}{spectra}(x, splice_at, interpolate_wvl = 5)
}
\arguments{
\item{x}{spectra object}

\item{splice_at}{wavelengths that serve as splice poits. Typically the
beginings of sensor 2 and sensor 3.}

\item{interpolate_wvl}{blah}
}
\value{
spectra object
}
\description{
\code{match_sensors} scales reflectance values of sensors 1 (vis) and 3 (swir2)
}
\details{
splice_at has no default because sensor transition points vary bwteeen vendors
and individual instruments. It is an imporant parameter though, so you should
visually inspect your spectra before assigning it.
Typical values in our own individual instruments were:
SVC ~ c(990, 1900)
PSR ~ c()
ASD ~ c()
}
\section{Methods (by class)}{
\itemize{
\item \code{spectra}: Match sensor ovelap regions
}}

\author{
Jose Eduardo Meireles and Anna Schweiger
}
