% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_sensors.R
\name{match_sensors}
\alias{match_sensors}
\alias{match_sensors.spectra}
\title{Match spectra at sensor transitions}
\usage{
match_sensors(x, splice_at, interpolate_wvl = 5, factor_range = c(0.5, 2))

\method{match_sensors}{spectra}(x, splice_at, interpolate_wvl = 5,
  factor_range = c(0.5, 2))
}
\arguments{
\item{x}{spectra object}

\item{splice_at}{wavelengths that serve as splice points Typically the
beginnings of sensor 2 and sensor 3.}

\item{interpolate_wvl}{blah}

\item{factor_range}{range of acceptable correction factors (min, max).
Defaults to c(0.5, 2)}
}
\value{
spectra object
}
\description{
\code{match_sensors} scales reflectance values of sensors 1 (vis) and 3 (swir2)
}
\details{
splice_at has no default because sensor transition points vary between vendors
and individual instruments. It is an important parameter though, so you should
visually inspect your spectra before assigning it.
Typical values in our own individual instruments were:
SVC ~ c(990, 1900)
PSR ~ c()
ASD ~ c(1001, 1801)

If the factors used to match spectra are unreasonable, \code{match_sensors}
will throw. Unreasonable factors (f) are defined as 0.5 > f < 1.5 or NaN,
which  happens when the reflectance value for the right sensor is 0.
}
\section{Methods (by class)}{
\itemize{
\item \code{spectra}: Match sensor overlap regions
}}

\author{
Jose Eduardo Meireles and Anna Schweiger
}
