% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_summary_overloads.R
\name{max.spectra}
\alias{max.spectra}
\title{Maximum reflectance}
\usage{
\method{max}{spectra}(..., na.rm = FALSE)
}
\arguments{
\item{...}{spectra object}

\item{na.rm}{boolean. remove NAs? Defaults to FALSE}
}
\value{
single numeric value
}
\description{
\code{max} Returns the maximum reflectance value in a spectra object
}
\author{
Jose Eduardo Meireles
}

