% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{meta<-}
\alias{meta<-}
\title{Set metadata}
\usage{
meta(x, label, sample) <- value
}
\arguments{
\item{x}{spectra object (lhs)}

\item{label}{metadata column label}

\item{sample}{sample name}

\item{value}{rhs. TODO}
}
\value{
nothing. called for its side effect
}
\description{
\code{meta} sets metadata
}
\author{
meireles
}

