% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{meta}
\alias{meta}
\alias{meta.spectra}
\title{Get metadata}
\usage{
meta(x, i, k, simplify = TRUE)

\method{meta}{spectra}(x, i, k, simplify = TRUE)
}
\arguments{
\item{x}{spectra object}

\item{i}{sample index or names}

\item{k}{metadata column index or names}

\item{simplify}{boolean. defaults to TRUE}
}
\value{
TODO
}
\description{
\code{meta} returns metadata of spectra
}
\section{Methods (by class)}{
\itemize{
\item \code{spectra}: get metadata
}}
\author{
meireles
}

