% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{names<-.spectra}
\alias{names<-.spectra}
\title{Set sample names in spectra}
\usage{
\method{names}{spectra}(spec) <- value
}
\arguments{
\item{spec}{spectra object to have their sample names modified}
}
\value{
nothing. called for its side effect.
}
\description{
Set sample names in spectra
}

