% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{names.spectra}
\alias{names.spectra}
\title{Get spectra sample names}
\usage{
\method{names}{spectra}(x)
}
\arguments{
\item{x}{spectra object}
}
\value{
vector of sample names
}
\description{
\code{names} returns a vector of sample names
}
\examples{
library(spectrolab)
spec = as.spectra(spec_matrix_example, name_idx = 1)
names(spec)
}
\author{
Jose Eduardo Meireles
}
