% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_spectra.R
\name{normalize_spectra}
\alias{normalize_spectra}
\title{Vector normalize spectra}
\usage{
normalize_spectra(x)
}
\arguments{
\item{x}{spectra object to be vector normalized}
}
\value{
spectra object with normalized spectra
}
\description{
\code{normalize_spectra} returns a spectra obj with vector normalized reflectances
}

