% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_operator_overloads.R
\name{Ops.spectra}
\alias{Ops.spectra}
\title{Operator overloading for spectra}
\usage{
\method{Ops}{spectra}(e1, e2)
}
\arguments{
\item{e1}{lhs}

\item{e2}{rhs}
}
\value{
operators on spectra
}
\description{
Operator overloading for spectra
}

