% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_operator_overloads.R
\name{Ops.spectra}
\alias{Ops.spectra}
\title{Arithmetic operators for spectra}
\usage{
\method{Ops}{spectra}(e1, e2)
}
\arguments{
\item{e1}{lhs}

\item{e2}{rhs}
}
\value{
Depends on the operator. c("+", "-", "*", "/", "^") return spectra
        c("==", "!=", "<", "<=", ">=", ">") return boolean matrices
}
\description{
Overloads arithmetic operators for spectra using `Ops.`
}
\author{
Jose Eduardo Meireles
}
