% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot.spectra}
\alias{plot.spectra}
\title{Plot spectra}
\usage{
\method{plot}{spectra}(spec, ylab = "Reflectance", xlab = "Wavelength",
  col = "black", lty = 2, type = "l", ...)
}
\arguments{
\item{spec}{spectra object}

\item{ylab}{label for y axis. defaults to "Reflectance"}

\item{xlab}{label for x axis. defaults to "Wavelength"}

\item{col}{line color. defaults to "black"}

\item{lty}{line type. defaults to 2}

\item{...}{other arguments passed to matplot}
}
\value{
nothing. called for side effect
}
\description{
Plot spectra
}

