% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot.spectra}
\alias{plot.spectra}
\title{Plot spectra}
\usage{
\method{plot}{spectra}(x, ylab = "Reflectance", xlab = "Wavelength",
  col = "black", lty = 1, type = "l", ...)
}
\arguments{
\item{x}{spectra object}

\item{ylab}{label for y axis. defaults to "Reflectance"}

\item{xlab}{label for x axis. defaults to "Wavelength"}

\item{col}{line color. defaults to "black"}

\item{lty}{line type. defaults to 1}

\item{type}{type of plot. meant to take either line "l" or no plotting "n"}

\item{...}{other arguments passed to matplot}
}
\value{
nothing. called for side effect
}
\description{
Plot spectra
}
\author{
meireles
}

