% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{plot_interact}
\alias{plot_interact}
\title{Interactive spectra plot}
\usage{
plot_interact(spec)
}
\arguments{
\item{spec}{spectra object}
}
\value{
interactive plot
}
\description{
\code{plot_interact} iteratively plots a defined number of spectra. This is
helpful for data inspection and large datasets.
}
\examples{
\dontrun{
# Create a reflectance matrix.
rf <- spec_matrix_example[, -1]

# Create a vector with wavelength labels.
wl <- colnames(rf)

# Sample with replacement to increase the number of spectra.
rf <- rf[sample(nrow(rf),300,replace = T),]

# Create a vector with sample labels that match the reflectance matrix rows.
sn <- paste(rep("sp",300), 1:300, sep = "_")

# Construct the spectra object using the `spectra` constructor
spec <- spectra(reflectance = rf, wavelengths = wl, names = sn)

# Start interactive plot
plot_interact(spec)
}
}
\author{
Anna K. Schweiger
}
