% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_quantile}
\alias{plot_quantile}
\title{Plot polygon for spectra quantiles}
\usage{
plot_quantile(spec, total_prob = 0.95, col = rgb(0, 0, 0, 0.1),
  add = FALSE, ...)
}
\arguments{
\item{spec}{Spectra object}

\item{total_prob}{Total mass to encompass. Number between 0.0 and 1.0}

\item{col}{Polygon color}

\item{add}{If add = FALSE (default), a new plot is created. Otherwise
(add = T), the quantile is added to the current plot.}

\item{...}{Other parameters passed to polygon()}
}
\value{
nothing. called for its side effect
}
\description{
Plot polygon for spectra quantiles
}

