% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_quantile}
\alias{plot_quantile}
\title{Plot spectra quantiles}
\usage{
plot_quantile(spec, total_prob = 0.95, col = rgb(0, 0, 0, 0.1),
  border = TRUE, add = FALSE, na.rm = FALSE, ...)
}
\arguments{
\item{spec}{spectra object}

\item{total_prob}{total probability mass to encompass. Single number
between 0.0 and 1.0. Defaults to 0.95.}

\item{col}{polygon color}

\item{border}{boolean. Draw border?}

\item{add}{if add = FALSE (default), a new plot is created. Otherwise
(add = TRUE), the quantile is added to the current plot.}

\item{na.rm}{boolean. remove NAs before computing the quantiles?}

\item{...}{other parameters passed to polygon() or to plot.}
}
\value{
nothing. Called for its side effect.
}
\description{
\code{plot_quantile} plots polygons for the quantiles of spectra per wavelength.
}
\author{
Jose Eduardo Meireles
}
