% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_SVC.R
\name{preprocess}
\alias{clip_spec}
\alias{preprocess}
\title{Preprocessing}
\usage{
clip_spec(x)
}
\arguments{
\item{x}{A data frame containing reflectance values.}
}
\value{
The output is a dataframe.
}
\description{
Function to read .sig files and combine them in a data.frame
}

