% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_SVC.R
\name{preprocess}
\alias{exclbadstart_spec}
\alias{excldipnir_spec}
\alias{excldipswir_spec}
\alias{excldropnir_spec}
\alias{exclhilo_spec}
\alias{exclhired_spec}
\alias{exclhistart_spec}
\alias{exclhivis_spec}
\alias{exclmaxvisshift_spec}
\alias{exclminidipnir_spec}
\alias{exclups_spec}
\alias{exclvisnirjump_spec}
\alias{preprocess}
\title{Preprocessing}
\usage{
exclhilo_spec(x)

exclhivis_spec(x)

exclhistart_spec(x)

exclbadstart_spec(x)

excldipnir_spec(x)

exclminidipnir_spec(x)

exclups_spec(x)

excldipswir_spec(x)

exclhired_spec(x)

exclmaxvisshift_spec(x)

excldropnir_spec(x)

exclvisnirjump_spec(x)
}
\arguments{
\item{x}{A data frame containing reflectance values.}
}
\value{
The output is a dataframe.
}
\description{
Function to read .sig files and combine them in a data.frame
}

