% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_summary_overloads.R
\name{print.spectra}
\alias{print.spectra}
\title{Print spectra}
\usage{
\method{print}{spectra}(x)
}
\arguments{
\item{x}{spectra object}
}
\value{
nothing. called for side effect
}
\description{
\code{print} prints to the console basic information about the spectra obj
}

