% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_summary_overloads.R
\name{print.spectra}
\alias{print.spectra}
\title{Prints spectra object}
\usage{
\method{print}{spectra}(spec)
}
\arguments{
\item{spec}{spectra object}
}
\value{
nothing. called for side effect
}
\description{
Prints spectra object
}

