% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_summary_overloads.R
\name{quantile.spectra}
\alias{quantile.spectra}
\title{Compute spectra quantiles by wavelength}
\usage{
\method{quantile}{spectra}(spec, probs = c(0.025, 0.25, 0.5, 0.75, 0.975))
}
\arguments{
\item{spec}{spectra object}

\item{probs}{Probabilities to compute quantiles.
Must be a vector of numerics between 0.0 and 1.0.
Defaults to c(0.025, 0.25, 0.5, 0.75, 0.975)}
}
\value{
spectra object with one spectrum for each prob
}
\description{
Compute spectra quantiles by wavelength
}

