% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ratio}
\alias{ratio}
\alias{ratio.spectra}
\title{Pairwise reflectance ratios}
\usage{
ratio(x, simplify = FALSE)

\method{ratio}{spectra}(x, simplify = FALSE)
}
\arguments{
\item{x}{spectra}

\item{simplify}{coerce to matrix or keep result as list}
}
\value{
list or matrix
}
\description{
\code{ratio} computes pairwise ratios between bands
}
\section{Methods (by class)}{
\itemize{
\item \code{spectra}: Compute pairwise reflectance ratios
}}

\examples{
library(spectrolab)
spec = as.spectra(spec_matrix_example)
spec_ratio_mat = ratio(spec)
}
\author{
Jose Eduardo Meireles
}
