% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectra.R
\name{read_spectra}
\alias{read_spectra}
\title{Read files from various formats into `spectra`}
\usage{
read_spectra(path, format, include_white_ref = FALSE, recursive = FALSE,
  exclude_if_matches = NULL, outside_01_action = "nothing", ...)
}
\arguments{
\item{path}{Path to directory or input files}

\item{format}{file formats. Currently the only option is "sig" or "svc" (for SVC).}

\item{include_white_ref}{boolean. include white reference. NOT IMPLEMENTED YET}

\item{recursive}{read files recursively}

\item{exclude_if_matches}{excludes files that match this regular expression.
Example: "BAD"}

\item{outside_01_action}{what to do with values outside 0 and 1? Options are
"NA", which replaces those values with NA or
"nothing" (default).}

\item{...}{nothing yet}
}
\value{
a single `spectra` or a list of `spectra` (in case files had diff
        number of wavelengths)
}
\description{
Read files from various formats into `spectra`
}
\author{
meireles
}

