% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{reflectance<-}
\alias{reflectance<-}
\title{Set reflectance in spectra}
\usage{
reflectance(spec) <- value
}
\arguments{
\item{spec}{Spectra object}
}
\value{
nothing. deleted function
}
\description{
Set reflectance in spectra
}

