% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{reflectance<-}
\alias{reflectance<-}
\title{Set spectra reflectance}
\usage{
reflectance(x) <- value
}
\arguments{
\item{x}{spectra object}

\item{value}{value to be assigned to the lhs}
}
\value{
nothing. called for its side effect
}
\description{
\code{reflectance} Assigns the rhs to the reflectance of the lhs spectra obj
}
\examples{
library(spectrolab)
spec = as.spectra(spec_matrix_example)
# scale all refletance values by 2
reflectance(spec) = reflectance(spec) * 2
}
\author{
Jose Eduardo Meireles
}
