% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{reflectance}
\alias{reflectance}
\title{Get the reflectance from spectra}
\usage{
reflectance(spec)
}
\arguments{
\item{spec}{spectra object}
}
\value{
matrix with samples in rows and wavelengths in columns
}
\description{
Get the reflectance from spectra
}

