% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{reflectance}
\alias{reflectance}
\alias{reflectance.spectra}
\title{Get spectra reflectance}
\usage{
reflectance(x)

\method{reflectance}{spectra}(x)
}
\arguments{
\item{x}{spectra object}
}
\value{
matrix with samples in rows and wavelengths in columns
}
\description{
\code{reflectance} returns the reflectance matrix from spectra
}
\section{Methods (by class)}{
\itemize{
\item \code{spectra}: Get spectra reflectance
}}

\examples{
library(spectrolab)
spec = as.spectra(spec_matrix_example, name_idx = 1)
is.matrix(reflectance(spec))
}
\author{
Jose Eduardo Meireles
}
