% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spectra.R
\name{resample}
\alias{resample}
\alias{resample.spectra}
\title{Resample spectra}
\usage{
resample(x, new_wvls, ...)

\method{resample}{spectra}(x, new_wvls, ...)
}
\arguments{
\item{x}{spectra object}

\item{new_wvls}{numeric vector of wavelengths to sample from spectra}

\item{...}{additional parameters passed to the \code{smooth.spline} function.}
}
\value{
spectra object with resampled spectra
}
\description{
\code{resample} returns spectra resampled to new wavelengths using smoothing.
Possible to increase or decrease the spectral resolution.
}
\section{Methods (by class)}{
\itemize{
\item \code{spectra}: Resample spectra
}}

