% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_function_overloads.R
\name{sd.spectra}
\alias{sd.spectra}
\title{Standard deviation spectrum}
\usage{
\method{sd}{spectra}(x, na.rm = TRUE)
}
\arguments{
\item{x}{spectra}

\item{na.rm}{boolean. remove NAs?}
}
\value{
single spectrum
}
\description{
Standard deviation spectrum
}
\author{
Jose Eduardo Meireles
}
