% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth.default}
\alias{smooth.default}
\title{Default smoothing function}
\usage{
\method{smooth}{default}(x, ...)
}
\arguments{
\item{x}{data to smooth over}

\item{...}{additional arguments}
}
\value{
smoothed data
}
\description{
Default smoothing function
}
