% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spectra.R
\name{smooth.default}
\alias{smooth.default}
\title{Smooth function from `stats`}
\usage{
\method{smooth}{default}(x, kind = c("3RS3R", "3RSS", "3RSR", "3R", "3", "S"),
  twiceit = FALSE, endrule = c("Tukey", "copy"), do.ends = FALSE)
}
\description{
Smooth function from `stats`
}

