% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spectra.R
\name{smooth}
\alias{smooth}
\alias{smooth.spectra}
\title{Smooth spectra}
\usage{
smooth(x, method = "spline", ...)

\method{smooth}{spectra}(x, method = "spline", ...)
}
\arguments{
\item{x}{spectra object. Wavelengths must be strictly increasing}

\item{method}{Choose smoothing method: "spline" (default) or "moving_average"}

\item{...}{additional parameters passed to \code{smooth.spline}.}
}
\value{
a spectra object of with smoothed spectra
}
\description{
\code{smooth} runs each spectrum by a smoothing and returns the spectra
}
\section{Methods (by class)}{
\itemize{
\item \code{spectra}: Smooth spectra
}}

\author{
Jose Eduardo Meireles
}
