% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_spline}
\alias{smooth_spline}
\title{Smooth spline functions for spectra}
\usage{
smooth_spline(x, parallel = TRUE, return_fn = FALSE, ...)
}
\arguments{
\item{x}{spectra object. bands must be strictly increasing}

\item{parallel}{boolean. Do computation in parallel? Defaults to TRUE.
Unfortunately, the parallelization does not work on Windows.}

\item{return_fn}{Boolean. If TRUE, \code{smooth_spline} returns the spline
functions instead of the smoothed spectra. Defaults to FALSE}

\item{...}{additional parameters passed to smooth.spline except nknots, which
is computed internally}
}
\value{
Smoothed spectra or, if return_fn = TRUE, a list of spline functions.
}
\description{
Gets spline functions for each spectrum in a spectra object.
}
\author{
Jose Eduardo Meireles
}
