% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{spectra}
\alias{spectra}
\title{Spectra object constructor}
\usage{
spectra(reflectance, wavelengths, names, meta = NULL, enforce01 = FALSE,
  ...)
}
\arguments{
\item{reflectance}{N by M numeric matrix. N samples in rows and M wavelengths
in columns}

\item{wavelengths}{wavelength names in vector of length M}

\item{names}{sample names in vector of length N}

\item{meta}{spectra metadata. defaults to NULL. Must be either of length or nrow
equals to the number of samples (nrow(reflectance) or length(names))}

\item{enforce01}{Force reflectance to be between 0 and 1. defaults to FALSE}

\item{...}{additional arguments to metadata creation. not implemented yet}
}
\value{
spectra object
}
\description{
\code{spectra} "manually" creates a spectra object
}
\note{
This function resorts to an ugly hack to deal with metadata assignment.
      Need to think a little harder to find a solution.
}
\author{
Jose Eduardo Meireles
}

