% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{spectra}
\alias{spectra}
\title{Create a spectra object}
\usage{
spectra(reflectance, wavelengths, names, meta = NULL, ...)
}
\arguments{
\item{reflectance}{N by M numeric matrix. N samples in rows. values between 0 and 1.}

\item{wavelengths}{wavelength names in vector of length M}

\item{names}{sample names in vector of length N}

\item{meta}{spectra metadata. defaults to NULL. Must be either of length or nrow
equals to the number of samples (i.e. nrow(reflectance) or length(names) )}
}
\value{
spectra object
}
\description{
Create a spectra object
}

