% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrolab.R
\docType{package}
\name{spectrolab-package}
\alias{spectrolab}
\alias{spectrolab-package}
\title{Spectrolab}
\description{
Class and methods for hyperspectral data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=spectrolab}
}

}
\author{
\strong{Maintainer}: Jose Eduardo Meireles \email{jemeireles@gmail.com}

Authors:
\itemize{
  \item Anna K. Schweiger \email{anna.k.schweiger@gmail.com}
  \item Jeannine Cavender-Bares \email{cavender@umn.edu}
}

}
\keyword{internal}
