% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{[<-.spectra}
\alias{[<-.spectra}
\title{Assign reflectance vlaues to spectra}
\usage{
\method{[}{spectra}(this, i, j) <- value
}
\arguments{
\item{i}{sample name}

\item{j}{wavelength}
}
\description{
Assign reflectance vlaues to spectra
}

