% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_operator_overloads.R
\name{t.spectra}
\alias{t.spectra}
\title{Spectra Transpose}
\usage{
\method{t}{spectra}(x)
}
\arguments{
\item{x}{spectra}
}
\value{
No return value. Operation not allowed
}
\description{
spectra are not transposable. Transpose the value instead
}
\examples{
library(spectrolab)
s = as_spectra(spec_matrix_example, name_idx = 1)

t(value(s))
t(as.matrix(s))
}
\author{
Jose Eduardo Meireles
}
