% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{wavelengths<-}
\alias{wavelengths<-}
\title{Set sample names in spectra}
\usage{
wavelengths(spec) <- value
}
\arguments{
\item{spec}{Spectra object}
}
\value{
nothing. called for its side effect.
}
\description{
Set sample names in spectra
}

