% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{wavelengths<-}
\alias{wavelengths<-}
\title{Set wavelength labels}
\usage{
wavelengths(x) <- value
}
\arguments{
\item{x}{spectra object (rhs)}

\item{value}{rhs}
}
\value{
nothing. called for its side effect.
}
\description{
\code{wavelengths} sets wavelength labels of lhs to the rhs values
}

