% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{wavelengths}
\alias{wavelengths}
\title{Get wavelength labels from spectra}
\usage{
wavelengths(spec)
}
\arguments{
\item{spec}{Spectra object}
}
\description{
Get wavelength labels from spectra
}

