% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{wavelengths}
\alias{wavelengths}
\title{Get wavelength labels from spectra}
\usage{
wavelengths(spec, return_num = TRUE)
}
\arguments{
\item{spec}{spectra object}

\item{return_num}{boolean. return vector of numeric values (default).
otherwise, a vector of strings is returned}
}
\value{
vector of wavelengths. numeric if `return_num` = TRUE (default).
}
\description{
Get wavelength labels from spectra
}

