% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{wavelengths}
\alias{wavelengths}
\alias{wavelengths.spectra}
\title{Get spectra wavelength labels}
\usage{
wavelengths(x, min = NULL, max = NULL, return_num = TRUE)

\method{wavelengths}{spectra}(x, min = NULL, max = NULL,
  return_num = TRUE)
}
\arguments{
\item{x}{spectra object}

\item{min}{= NULL}

\item{max}{= NULL}

\item{return_num}{boolean. return vector of numeric values (default).
otherwise, a vector of strings is returned}
}
\value{
vector of wavelengths. numeric if `return_num` = TRUE (default).
}
\description{
\code{wavelengths} returns a vector of wavelength labels from spectra
}
\section{Methods (by class)}{
\itemize{
\item \code{spectra}: Set spectra wavelength labels
}}

\examples{
library(spectrolab)
spec = as.spectra(spec_matrix_example, name_idx = 1)
head(wavelengths(spec))
}
\author{
Jose Eduardo Meireles
}
